feat: updating password and updating email support
Routes for updating the password and the email of the current authenticated user is now supported. Email update and password update ensure that they meet the system requirement and the email check ensures that its not already in use.

 * Both endpoints contain ApiDocs
 * email reset returns a verification Boolean expression marking if the user would require going through the email verification stage again. (only standard users).

+++ app/services/Verification.service.ts
@@ -16,6 +16,12 @@ export class VerificationService {
      * @param user The user who is getting their verification progress reset.
      */
     public static async reset(user: User) {
+        // If the given user is moderator or a admin, they should not be subject to updating user
+        // role state. e.g only standard users have to go through email verification again.
+        if (user.role === UserRole.ADMIN || user.role === UserRole.MODERATOR) {
+            return;
+        }
+
         const emailRepository = getCustomRepository(EmailVerificationRepository);
         await emailRepository.removeForUser(user);
